Add mutex lock on LA ID during transfer processing (#27323)
* Initial flow.

* Added trace code.

* Added check on job attempts.

* Increased timeout to 120 seconds.

* Removed check on attempts.

* Added RazorX.

* Reduced timeout.

* Replaced RazorX with feature flag.

* Added feature constant.

* Removed unnecessary if condition.

* Logged more data.

* Dummy commit.

+++ app/Jobs/TransferProcess.php
@@ -12,6 +12,8 @@ class TransferProcess extends Job
 {
     const MUTEX_LOCK_TIMEOUT = 600;
 
+    const JOB_RELEASE_DELAY = 5; // In seconds.
+
     protected $mutex;
 
     protected $repo;
@@ -19,6 +21,8 @@ class TransferProcess extends Job
     protected $payment;
 
     protected $transferMode;
+
+    protected $jobDeleteFlag;
 
     public $timeout = 900;
 
@@ -31,6 +35,8 @@ class TransferProcess extends Job
         $this->payment = $payment;
 
         $this->transferMode = $transfermode;
+
+        $this->jobDeleteFlag = true;
     }
 
     public function handle()
@@ -64,6 +70,24 @@ class TransferProcess extends Job
 
         }catch (\Exception $ex)
         {
+            if ($ex->getMessage() === Transfer\Constant::MUTEX_LOCK_ON_LINKED_ACCOUNT_ID_NOT_ACQUIRED)
+            {
+                $this->trace->traceException(
+                    $ex,
+                    null,
+                    TraceCode::TRANSFER_PROCESS_JOB_RELEASE_SINCE_MUTEX_ACQUIRE_FAILED,
+                    [
+                        'payment_id'    => $this->payment->getId(),
+                        'source_type'   => $this->transferMode,
+                        'job_attempts'  => $this->attempts(),
+                    ]
+                );
+
+                $this->jobDeleteFlag = false;
+
+                $this->release(self::JOB_RELEASE_DELAY);
+            }
+
             $this->trace->traceException(
                 $ex,
                 null,
@@ -77,7 +101,10 @@ class TransferProcess extends Job
         }
         finally
         {
-            $this->delete();
+            if ($this->jobDeleteFlag === true)
+            {
+                $this->delete();
+            }
         }
     }
